DjangoTeste/requirements.txt
Django==5.2.5
djangorestframework==3.16.0
openai==1.101.0
Requests==2.32.5
django-cors-headers>=4.7.0
pillow>=11.3.0
djangorestframework-simplejwt>=5.5.1
daphne>=4.2.1
channels-redis>=4.3.0
django-cleanup==9.0.0
django-allauth==65.11.2
django-htmx==1.24.1
faker==37.8.0